requirements.txt
requests~=2.28.2
python-socketio~=5.7.2
